Create js-homework-2b.html
Write a function that takes two number arguments - homeScore and awayScore - and returns a different string depending on the values you pass in. Assuming the two arguments represent final scores to a basketball game: if the home team wins, return 'The home team won'; if the away team wins, return 'The away team won.' In your code, call your function twice to make sure it's working - once for a time when the home team wins and once for when it loses. Each time you call your function, log the value it returns to the console.