Update js-homework-2b.html
Write a function that takes a single argument - someone's name. Write a function that: if the name that's passed in is someone in your group, return 'Hello, group member'; if the name is 'Andrew', return 'Hello, teacher'; otherwise, return 'Hello!' Call your function three times - once for each branch of the if statement that you wrote above and log the returned value to the console to demonstrate that it works. Use prompt() to ask visitors to your page for a name. Feed that name to your function and feed the message returned by the function to the console.